requirements.txt
requests
spotipy
pillow
pandas
tabulate